requirements.txt
PyPDF2>=3.0.0
python-docx>=1.1.0
google-genai>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
python-dotenv>=1.0.0
